+++ src/cli/index.ts
@@ -5,11 +5,12 @@ import { findEntryPoint, getPackageOption } from "./shared/index";
 import { start } from "./commands/start";
 import { startDevServer } from "./commands/dev";
 import { startTsDevServer } from "./commands/dev-ts";
+import generateAuthSecret from "./commands/gen-secret";
 
 const packageJson = getPackageOption();
 
 const cli = createCLI(
-  "Zoltra",
+  "zoltra",
   packageJson.version,
   "The official CLI for Zoltra - A fast, file-based JavaScript web server framework"
 );
@@ -41,6 +42,14 @@ cli
     const entryPoint = findEntryPoint();
     startTsDevServer(entryPoint);
   })
+  .finalize();
+
+cli
+  .command(
+    "gen-secret",
+    "Generate a new cryptographically secure secret for signing JWT tokens"
+  )
+  .action(generateAuthSecret)
   .finalize();
 
 cli.parse().catch(console.error);