+++ src/lib/core/context.ts
@@ -1,7 +1,7 @@
 import { IncomingMessage, ServerResponse } from "http";
 import { parse as parseQuery, ParsedUrlQuery } from "querystring";
 import { parse as parseUrl, UrlWithStringQuery } from "url";
-import { IApplication, IContainer, IContext, Registry } from "../../types";
+import { IApplication, IContainer, IContext } from "../../types";
 
 export class Context implements IContext {
   req: IncomingMessage;
@@ -171,7 +171,7 @@ export class Context implements IContext {
     return this._rawBody;
   }
 
-  service<T = any>(name: Registry.__services["name"]): T {
+  service<T = any>(name: string): T {
     return this.services.resolve<T>(name);
   }